requirements.txt
config==0.5.1
pandas==2.2.0
snowflake-connector-python==3.7.0